remove old cordovaDebugSession class
